Decouple submission side effects into independent jobs
UploadJob and ProcessSubmissionJob each performed a file copy, a
working-list sheet write, and two mailer deliveries sequentially.
A failure in the middle (e.g. the Google Sheets rate-limit incident
on 2026-04-18) would short-circuit later steps — most importantly
the curator notification mail — leaving the operator to notice and
manually recover.

Split each entry point into four fan-out operations that are
enqueued together and fail independently:

  - CopySubmissionFilesJob  (was the opening step of both jobs)
  - AddToWorkingListJob     (new, for the initial-submission flow)
  - UpdateWorkingListJob    (new, for the re-upload flow)
  - SubmissionMailer#submitter_confirmation / #curator_notification

UploadJob is renamed to CopySubmissionFilesJob and ProcessSubmissionJob
is removed entirely; both controllers now enqueue the four operations
directly. A single-point outage in any one operation no longer blocks
the others, and each can be retried or re-driven in isolation.
