+++ app/jobs/add_to_working_list_job.rb
@@ -1,0 +1,7 @@
+class AddToWorkingListJob < ApplicationJob
+  queue_as :default
+
+  def perform(submission)
+    WorkingList.instance.add_new_submission submission
+  end
+end
