+++ app/jobs/copy_submission_files_job.rb
@@ -1,0 +1,7 @@
+class CopySubmissionFilesJob < ApplicationJob
+  queue_as :default
+
+  def perform(upload)
+    upload.via.copy_files_to_submissions_dir
+  end
+end
