+++ app/jobs/update_working_list_job.rb
@@ -1,0 +1,7 @@
+class UpdateWorkingListJob < ApplicationJob
+  queue_as :default
+
+  def perform(submission)
+    WorkingList.instance.update_data_arrival_date submission
+  end
+end
