--- app/jobs/upload_job.rb
@@ -1,14 +1,0 @@
-class UploadJob < ApplicationJob
-  queue_as :default
-
-  def perform(upload)
-    upload.via.copy_files_to_submissions_dir
-
-    submission = upload.submission
-
-    WorkingList.instance.update_data_arrival_date submission
-
-    SubmissionMailer.with(submission:).submitter_confirmation.deliver_later
-    SubmissionMailer.with(submission:).curator_notification.deliver_later
-  end
-end
