+++ test/integration/submissions/uploads_test.rb
@@ -1,6 +1,8 @@
 require 'test_helper'
 
 class Submissions::UploadsTest < ActionDispatch::IntegrationTest
+  include ActionMailer::TestHelper
+
   setup do
     @user = users(:alice)
 
@@ -23,8 +25,14 @@ class Submissions::UploadsTest < ActionDispatch::IntegrationTest
     path  = Rails.application.config_for(:app).upload_events_log!
     event = JSON.parse(File.read(path).lines.last)
 
-    assert_equal submission.mass_id,        event['mass_id']
-    assert_equal @user.uid,                 event['dway_account']
-    assert_match(/\A\d{8}-\d{6}\z/,         event['data_arrival_date'])
+    assert_equal submission.mass_id, event['mass_id']
+    assert_equal @user.uid,          event['dway_account']
+    assert_match /\A\d{8}-\d{6}\z/,  event['data_arrival_date']
+
+    assert_enqueued_with job: CopySubmissionFilesJob
+    assert_enqueued_with job: UpdateWorkingListJob, args: [submission]
+
+    assert_enqueued_email_with SubmissionMailer, :submitter_confirmation, params: {submission:}
+    assert_enqueued_email_with SubmissionMailer, :curator_notification,   params: {submission:}
   end
 end
